+++ packages/at_secondary_server/lib/src/connection/connection_metrics.dart
@@ -1,11 +1,13 @@
-import 'package:at_secondary/src/connection/inbound/connection_util.dart';
 import 'package:at_secondary/src/server/at_secondary_impl.dart';
 import 'package:at_server_spec/at_server_spec.dart';
 
 class ConnectionMetricsImpl implements ConnectionMetrics {
   @override
   int getInboundConnections() {
-    return ConnectionUtil.getActiveConnectionSize();
+    return AtSecondaryServerImpl.getInstance()
+        .inboundConnectionManager
+        .pool
+        .getActiveConnectionSize();
   }
 
   @override