+++ packages/at_secondary_server/lib/src/connection/inbound/connection_util.dart
@@ -138,8 +138,8 @@ class InboundIdleChecker {
     // Unauthenticated connections should be reaped increasingly aggressively as we approach max connections
     // Authenticated connections should also be reaped as we approach max connections, but a lot less aggressively
     // Ultimately, the caller (e.g. [InboundConnectionManager] decides **whether** to reap or not.
-    int? poolMaxConnections = owningPool!.getCapacity();
-    int lowWaterMark = (poolMaxConnections! * lowWaterMarkRatio).floor();
+    int poolMaxConnections = owningPool!.getCapacity();
+    int lowWaterMark = (poolMaxConnections * lowWaterMarkRatio).floor();
     int numConnectionsOverLwm =
         max(owningPool!.getCurrentSize() - lowWaterMark, 0);
 
@@ -184,36 +184,5 @@ class InboundIdleChecker {
       var retVal = actualIdleTime > allowableIdleTime;
       return retVal;
     }
-  }
-}
-
-class ConnectionUtil {
-  /// Returns the number of active monitor connections.
-  static int getMonitorConnectionSize() {
-    var count = 0;
-    InboundConnectionPool.getInstance().getConnections().forEach((connection) {
-      if (!connection.isInValid() && connection.isMonitor!) {
-        count++;
-      }
-    });
-
-    return count;
-  }
-
-  /// Returns the number of active connections.
-  static int getActiveConnectionSize() {
-    var count = 0;
-    InboundConnectionPool.getInstance().getConnections().forEach((connection) {
-      if (!connection.isInValid()) {
-        count++;
-      }
-    });
-
-    return count;
-  }
-
-  /// Return total capacity of connection manager of connection pool.
-  static int totalConnectionSize() {
-    return InboundConnectionPool.getInstance().getConnections().length;
   }
 }